+++ src/common/app-state.ts
@@ -28,7 +28,8 @@ export type StateType = {
         texture2?:  string,
         fadeTrans?: number,
         fadeWait?:  number,
-        transition: number,
+        transType:  string,
+        transTime:  number,
         rotationZ:  number
     },
     monitor: {
@@ -209,7 +210,8 @@ export const StateSchema = `{
         texture2:   string,
         fadeTrans:  number,
         fadeWait:   number,
-        transition: number,
+        transType:  string,
+        transTime:  number,
         rotationZ:  number
     },
     monitor: {
@@ -386,7 +388,8 @@ export const StateDefault: StateType = {
         texture2:   "/canvas/STD/hexagons-2.jpg",
         fadeTrans:  150,
         fadeWait:   10000,
-        transition: 2.0,
+        transType:  "fade",
+        transTime:  2.0,
         rotationZ:  0
     },
     monitor: {